cleanup mythological code
adds a clearburrows command as a band-aid fix for frequent ghost mob
burrows - this should ideally be properly fixed better later, but I'm
not about to try to do that myself.

use the existing heightmap predicate instead of trying to roll our own
(I would've liked to use the existing code for this, but I don't think
Hypixel sends the data necessary for that?)

fix inquisitor waypoints never actually being created because of an
incorrect `is` check on `ServerPlayerEntity`, which is - in fact - not
used by the client on a multiplayer server.